+++ src/Sagas/Configuration/AnnotationsBasedSagaConfigurationLoader.php
@@ -13,10 +13,10 @@ declare(strict_types = 1);
 
 namespace Desperado\ServiceBus\Sagas\Configuration;
 
-use Desperado\ServiceBus\AnnotationsReader\Annotation;
-use Desperado\ServiceBus\AnnotationsReader\AnnotationCollection;
-use Desperado\ServiceBus\AnnotationsReader\AnnotationsReader;
-use Desperado\ServiceBus\AnnotationsReader\DefaultAnnotationsReader;
+use Desperado\ServiceBus\Infrastructure\AnnotationsReader\Annotation;
+use Desperado\ServiceBus\Infrastructure\AnnotationsReader\AnnotationCollection;
+use Desperado\ServiceBus\Infrastructure\AnnotationsReader\AnnotationsReader;
+use Desperado\ServiceBus\Infrastructure\AnnotationsReader\DefaultAnnotationsReader;
 use Desperado\ServiceBus\Common\Contract\Messages\Event;
 use Desperado\ServiceBus\MessageHandlers\Handler;
 use Desperado\ServiceBus\MessageHandlers\HandlerCollection;
@@ -253,7 +253,7 @@ final class AnnotationsBasedSagaConfigurationLoader implements SagaConfiguration
     {
         foreach($annotationCollection->classLevelAnnotations() as $annotation)
         {
-            /** @var \Desperado\ServiceBus\AnnotationsReader\Annotation $annotation */
+            /** @var \Desperado\ServiceBus\Infrastructure\AnnotationsReader\Annotation $annotation */
 
             $annotationObject = $annotation->annotationObject();